feat(rope): Centralize RoPE logic and fix precision issue
